+++ src/db/attmap.h
@@ -10,7 +10,7 @@ typedef struct Attribute
 {
 	int8_t *name;
 	DataType type;
-	uint16_t pos;
+	uint32_t pos;
 	struct Attribute *nxt, *nxt_sq;
 } Attribute;
 
@@ -18,15 +18,15 @@ typedef struct AttMap
 {
 	Attribute **map;
 	Attribute *head, *tail;
-	uint16_t tot_len;
+	uint32_t tot_len;
 } AttMap;
 
 AttMap *attmap_init();
 void attmap_deinit(AttMap *);
 void attmap_add_att(AttMap *, int8_t *, DataType,
-			uint16_t);
+			uint32_t);
 Attribute *attmap_find(AttMap *, int8_t *,
-			uint16_t *);
-uint16_t attribute_get_len(AttMap *, Attribute *);
+			uint32_t *);
+uint32_t attribute_get_len(AttMap *, Attribute *);
 
 #endif
